fix(dreaming): authenticate heartbeat transcript turns
