+++ src/plugin-sdk/memory-core-engine-runtime.ts
@@ -24,7 +24,6 @@ export type DreamingArtifactsAuditIssue = {
   severity: "warn" | "error";
   code:
     | "dreaming-session-corpus-unreadable"
-    | "dreaming-session-corpus-heartbeat-derived"
     | "dreaming-session-corpus-self-ingested"
     | "dreaming-session-ingestion-unreadable"
     | "dreaming-diary-unreadable";
@@ -37,8 +36,6 @@ export type DreamingArtifactsAuditSummary = {
   dreamsPath?: string;
   sessionCorpusDir: string;
   sessionCorpusFileCount: number;
-  heartbeatContaminatedSessionCorpusFileCount?: number;
-  heartbeatContaminatedSessionCorpusLineCount?: number;
   suspiciousSessionCorpusFileCount: number;
   suspiciousSessionCorpusLineCount: number;
   sessionIngestionPath: string;
@@ -54,8 +51,6 @@ export type RepairDreamingArtifactsResult = {
   archivedSessionCorpus: boolean;
   archivedSessionIngestion: boolean;
   archivedPaths: string[];
-  removedHeartbeatDerivedLines?: number;
-  clearedSessionCheckpointKeys?: number;
   warnings: string[];
 };
